refactor: 업로드 횟수 증가 로직 인터페이스 제거
